feat: Refactor out cx-address-card component
