+++ projects/storefrontapp-e2e-cypress/cypress/helpers/address-book.ts
@@ -25,26 +25,26 @@ export const assertAddressForm = (
   state?: string
 ): void => {
   state = state ? state : 'CA-QC';
-  cy.get('cx-address-card .card-header').contains('✓ DEFAULT');
-  cy.get('cx-address-card .card-body').within(_ => {
-    cy.get('.cx-address-card-label-name').should(
+  cy.get('cx-card .card-header').contains('✓ DEFAULT');
+  cy.get('cx-card .card-body').within(_ => {
+    cy.get('.cx-card-label-bold').should(
       'contain',
       `${address.firstName} ${address.lastName}`
     );
-    cy.get('.cx-address-card-label')
-      .first()
+    cy.get('.cx-card-label')
+      .eq(0)
       .should('contain', address.address.line1);
-    cy.get('.cx-address-card-label')
-      .next()
+    cy.get('.cx-card-label')
+      .eq(1)
       .should('contain', address.address.line2);
-    cy.get('.cx-address-card-label')
-      .next()
+    cy.get('.cx-card-label')
+      .eq(2)
       .should('contain', `${address.address.city}, ${state}`);
-    cy.get('.cx-address-card-label')
-      .next()
+    cy.get('.cx-card-label')
+      .eq(3)
       .should('contain', address.address.postal);
-    cy.get('.cx-address-card-label')
-      .next()
+    cy.get('.cx-card-label')
+      .eq(4)
       .should('contain', address.phone);
   });
 };
@@ -63,12 +63,14 @@ export function createNewAddress() {
 }
 
 export function verifyNewAddress() {
-  cy.get('cx-address-card').should('have.length', 1);
+  cy.get('cx-card').should('have.length', 1);
   assertAddressForm(newAddress);
 }
 
 export function editAddress() {
-  cy.get('.edit').click();
+  cy.get('a')
+    .contains('Edit')
+    .click();
   cy.get('cx-address-form').within(() => {
     cy.get('[formcontrolname="titleCode"]').ngSelect('Mr.');
     cy.get('[formcontrolname="firstName"]')
@@ -86,7 +88,7 @@ export function editAddress() {
 }
 
 export function verifyEditedAddress() {
-  cy.get('cx-address-card').should('have.length', 1);
+  cy.get('cx-card').should('have.length', 1);
   assertAddressForm(editedAddress);
 }
 
@@ -100,21 +102,26 @@ export function addSecondAddress() {
     .contains(' Add new address ')
     .click({ force: true });
   fillShippingAddress(secondAddress);
-  cy.get('cx-address-card').should('have.length', 2);
+  cy.get('cx-card').should('have.length', 2);
 }
 
 export function setSecondAddressToDefault() {
-  cy.get('.set-default').click();
+  cy.get('a')
+    .contains('Set as default')
+    .click();
 
-  const firstCard = cy.get('cx-address-card').first();
+  const firstCard = cy.get('cx-card').first();
   firstCard.should('contain', '✓ DEFAULT');
   firstCard.should('contain', 'N Z');
 }
 
 export function deleteExistingAddress() {
-  let firstCard = cy.get('cx-address-card').first();
+  const firstCard = cy.get('cx-card').first();
 
-  firstCard.find('.delete').click();
+  firstCard
+    .get('a')
+    .contains('Delete')
+    .click();
   cy.get('.cx-address-card-delete-msg').should(
     'contain',
     'Are you sure you want to delete this address?'
