+++ projects/storefrontlib/src/cms-components/myaccount/address-book/address-book.component.html
@@ -24,10 +24,14 @@
         *ngFor="let address of addresses$ | async"
         class="col-md-6 cx-address-card"
       >
-        <cx-address-card
-          (editEvent)="editAddressButtonHandle(address)"
-          [address]="address"
-        ></cx-address-card>
+        <cx-card
+          [border]="true"
+          [fitToContainer]="true"
+          [content]="getCardContent(address) | async"
+          (editCard)="editAddressButtonHandle(address)"
+          (setDefaultCard)="setAddressAsDefault(address.id)"
+          (deleteCard)="deleteAddress(address.id)"
+        ></cx-card>
       </div>
     </div>
   </ng-container>
