+++ projects/storefrontlib/src/cms-components/myaccount/address-book/address-book.component.ts
@@ -1,7 +1,14 @@
 import { Component, OnInit } from '@angular/core';
-import { Address } from '@spartacus/core';
-import { Observable } from 'rxjs';
+import {
+  Address,
+  TranslationService,
+  UserAddressService,
+  CheckoutDeliveryService,
+} from '@spartacus/core';
+import { Observable, combineLatest } from 'rxjs';
 import { AddressBookComponentService } from './address-book.component.service';
+import { Card } from '@spartacus/storefront';
+import { map } from 'rxjs/operators';
 
 @Component({
   selector: 'cx-address-book',
@@ -9,13 +16,19 @@ import { AddressBookComponentService } from './address-book.component.service';
 })
 export class AddressBookComponent implements OnInit {
   addresses$: Observable<Address[]>;
+  cards$: Observable<Card[]>;
   addressesStateLoading$: Observable<boolean>;
   currentAddress: Address;
 
   showAddAddressForm = false;
   showEditAddressForm = false;
 
-  constructor(public service: AddressBookComponentService) {}
+  constructor(
+    public service: AddressBookComponentService,
+    private translation: TranslationService,
+    private userAddressService: UserAddressService,
+    protected checkoutDeliveryService: CheckoutDeliveryService
+  ) {}
 
   ngOnInit(): void {
     this.addresses$ = this.service.getAddresses();
@@ -50,5 +63,52 @@ export class AddressBookComponent implements OnInit {
 
   editAddressCancel(): void {
     this.showEditAddressForm = false;
+  }
+
+  getCardContent(address: Address) {
+    return combineLatest([
+      this.translation.translate('addressCard.default'),
+      this.translation.translate('addressCard.setAsDefault'),
+      this.translation.translate('common.delete'),
+      this.translation.translate('common.edit'),
+    ]).pipe(
+      map(([defaultText, setAsDefaultText, textDelete, textEdit]) => {
+        let region = '';
+
+        if (address.region && address.region.isocode) {
+          region = address.region.isocode + ', ';
+        }
+
+        const actions: { name: string; event: string }[] = [];
+        if (!address.defaultAddress) {
+          actions.push({ name: setAsDefaultText, event: 'default' });
+        }
+        actions.push({ name: textEdit, event: 'edit' });
+        actions.push({ name: textDelete, event: 'delete' });
+
+        return {
+          textBold: address.firstName + ' ' + address.lastName,
+          text: [
+            address.line1,
+            address.line2,
+            address.town + ', ' + region + address.country.isocode,
+            address.postalCode,
+            address.phone,
+          ],
+          actions: actions,
+          header: address.defaultAddress ? `✓ ${defaultText}` : '',
+        };
+      })
+    );
+  }
+
+  setAddressAsDefault(addressId: string): void {
+    this.userAddressService.setAddressAsDefault(addressId);
+    this.checkoutDeliveryService.clearCheckoutDeliveryDetails();
+  }
+
+  deleteAddress(addressId: string): void {
+    this.userAddressService.deleteUserAddress(addressId);
+    this.checkoutDeliveryService.clearCheckoutDeliveryDetails();
   }
 }
