--- projects/storefrontlib/src/cms-components/myaccount/address-book/address-card/address-card.component.html
@@ -1,66 +1,0 @@
-<div class="card">
-  <div class="card-header" *ngIf="address?.defaultAddress && !editMode">
-    &#10003; {{ 'addressCard.default' | cxTranslate }}
-  </div>
-  <div
-    class="card-body cx-card-body"
-    [class.cx-address-card-delete-mode]="editMode"
-  >
-    <div *ngIf="editMode" class="cx-address-card-delete-msg">
-      {{ 'addressBook.areYouSureToDeleteAddress' | cxTranslate }}
-    </div>
-    <div class="cx-address-data">
-      <div class="cx-address-card-label-name">
-        {{ address?.firstName }} {{ address?.lastName }}
-      </div>
-      <div class="cx-address-card-label">{{ address?.line1 }}</div>
-      <div class="cx-address-card-label">{{ address?.line2 }}</div>
-      <div class="cx-address-card-label">
-        {{ address?.town }},
-        <span *ngIf="!address?.region?.isocode">{{
-          address?.country?.isocode
-        }}</span
-        ><span>{{ address?.region?.isocode }}</span>
-      </div>
-      <div class="cx-address-card-label">{{ address?.postalCode }}</div>
-      <div class="cx-address-card-label">{{ address?.phone }}</div>
-    </div>
-
-    <div *ngIf="editMode" class="row cx-address-card-delete">
-      <div class="col-md-6">
-        <button class="btn btn-block btn-secondary" (click)="cancelEdit()">
-          {{ 'common.cancel' | cxTranslate }}
-        </button>
-      </div>
-      <div class="col-md-6">
-        <button
-          (click)="deleteAddress(address.id)"
-          class="btn btn-block btn-primary"
-        >
-          {{ 'common.delete' | cxTranslate }}
-        </button>
-      </div>
-    </div>
-    <div *ngIf="!editMode" class="card-actions">
-      <a
-        *ngIf="!address?.defaultAddress"
-        (click)="setAddressAsDefault(address.id)"
-        class="card-link btn-link set-default"
-        tabindex="0"
-        >{{ 'addressCard.setAsDefault' | cxTranslate }}</a
-      >
-      <a
-        (click)="openEditFormEvent()"
-        class="card-link btn-link edit"
-        tabindex="0"
-        >{{ 'common.edit' | cxTranslate }}</a
-      >
-      <a
-        (click)="setEditMode()"
-        class="card-link btn-link delete"
-        tabindex="0"
-        >{{ 'common.delete' | cxTranslate }}</a
-      >
-    </div>
-  </div>
-</div>
