+++ projects/storefrontlib/src/cms-components/myaccount/address-book/index.ts
@@ -1,4 +1,3 @@
 export * from './address-book.component';
 export * from './address-book.component.service';
 export * from './address-book.module';
-export * from './address-card/address-card.component';
