+++ projects/storefrontstyles/scss/components/myaccount/_address-book.scss
@@ -7,10 +7,6 @@
     padding: 5rem 0 5rem 0;
   }
 
-  .cx-address-card {
-    padding-bottom: 30px;
-  }
-
   .cx-form {
     padding-top: 30px;
   }
