--- projects/storefrontstyles/scss/components/myaccount/_address-card.scss
@@ -1,35 +1,0 @@
-%cx-address-card {
-  .card {
-    margin: 0 0 1.25rem 0;
-    height: 100%;
-  }
-
-  .card-actions {
-    display: flex;
-    justify-content: flex-end;
-    padding: 1.25rem 0 0 0;
-  }
-
-  .cx-address-card-delete {
-    padding: 1.25rem 0 0 0;
-  }
-
-  .cx-address-card-label-bold {
-    @include weight('bold');
-  }
-
-  .cx-address-card-delete-msg {
-    color: var(--cx-color-danger);
-    padding: 0 0 1.25rem 0;
-  }
-
-  .cx-address-card-delete-mode {
-    background-color: var(--cx-color-background);
-  }
-
-  .cx-card-body {
-    justify-content: space-between;
-    display: flex;
-    flex-direction: column;
-  }
-}
