+++ projects/storefrontstyles/scss/components/myaccount/_index.scss
@@ -1,7 +1,6 @@
 @import './anonymous-consent/index';
 @import './order-history/index';
 @import './address-book';
-@import './address-card';
 @import './payment-methods';
 @import './close-account';
 @import './close-account-modal';
@@ -15,16 +14,16 @@
 @import './anonymous-consent/index';
 @import './my-interests';
 
-$myaccount-components-whitelist: cx-address-book, cx-address-card,
-  cx-order-summary, cx-order-history, cx-order-confirmation-items,
-  cx-order-confirmation-totals, cx-order-confirmation-overview,
-  cx-order-confirmation-thank-you-message, cx-promotions,
-  cx-order-detail-headline, cx-order-detail-items, cx-order-detail-shipping,
-  cx-order-detail-totals, cx-order-detail-totals, cx-close-account,
-  cx-close-account-modal, cx-anonymous-consent-management-banner,
-  cx-anonymous-consent-dialog, cx-anonymous-consent-open-dialog,
-  cx-consent-management-form, cx-consent-management, cx-forgot-password,
-  cx-tracking-events, cx-order-return-request-list, cx-amend-order-actions,
+$myaccount-components-whitelist: cx-address-book, cx-order-summary,
+  cx-order-history, cx-order-confirmation-items, cx-order-confirmation-totals,
+  cx-order-confirmation-overview, cx-order-confirmation-thank-you-message,
+  cx-promotions, cx-order-detail-headline, cx-order-detail-items,
+  cx-order-detail-shipping, cx-order-detail-totals, cx-order-detail-totals,
+  cx-close-account, cx-close-account-modal,
+  cx-anonymous-consent-management-banner, cx-anonymous-consent-dialog,
+  cx-anonymous-consent-open-dialog, cx-consent-management-form,
+  cx-consent-management, cx-forgot-password, cx-tracking-events,
+  cx-order-return-request-list, cx-amend-order-actions,
   cx-return-order-confirmation, cx-cancel-order-confirmation, cx-my-interests,
   cx-amend-order-items, cx-order-details-actions, cx-consignment-tracking,
   cx-return-request-overview, cx-return-request-totals, cx-return-request-items,
